refactor(core): reorganize auth actions
