+++ packages/core/src/actions/index.ts
@@ -1,12 +1,12 @@
 export { signInAction } from "@/actions/signIn/signIn.ts"
-export { signInCredentialsAction } from "@/actions/signInCredentials/signInCredentials.ts"
+export { signInCredentialsAction } from "@/actions/signIn/signInCredentials.ts"
 export { callbackAction } from "@/actions/callback/callback.ts"
 export { sessionAction } from "@/actions/session/session.ts"
 export { signOutAction } from "@/actions/signOut/signOut.ts"
 export { csrfTokenAction } from "@/actions/csrfToken/csrfToken.ts"
-export { updateSessionAction } from "@/actions/updateSession/updateSession.ts"
+export { updateSessionAction } from "@/actions/session/updateSession.ts"
 export { signUpAction } from "@/actions/signUp/signUp.ts"
-export { tokensAction } from "@/actions/tokens/getProviderTokens.ts"
-export { revokeAction } from "@/actions/tokens/revoke.ts"
-export { refreshAction } from "@/actions/user/refresh.ts"
+export { tokensAction } from "@/actions/providers/tokens/tokens.ts"
+export { revokeAction } from "@/actions/providers/tokens/revoke.ts"
+export { refreshAction } from "@/actions/providers/user/refresh.ts"
 export { disconnectAction } from "@/actions/providers/disconnect.ts"
