--- packages/core/src/actions/oidc/index.ts
@@ -1,10 +1,0 @@
-export { discoveryMetadata, DISCOVERY_SUFFIX, normalizeIssuer } from "@/actions/oidc/discovery.ts"
-export {
-    resolveOpenIDProvider,
-    isOIDCProvider,
-    createOpenIDPlaceholder,
-    clearResolvedProviderCache,
-} from "@/actions/oidc/resolve-provider.ts"
-export { fetchJWKS, getJWKSVerifier, ensureJWKSValidated, clearJWKSCache } from "@/actions/oidc/jwks.ts"
-export { validateIDToken, type ValidateIDTokenOptions } from "@/actions/oidc/id-token.ts"
-export { createOIDCAuthorizationURL } from "@/actions/oidc/authorization-url.ts"
