+++ packages/core/src/actions/session/updateSession.ts
