+++ packages/core/src/api/signIn.ts
@@ -1,10 +1,10 @@
-import { cacheControl, secureApiHeaders } from "@/shared/headers.ts"
 import { HeadersBuilder } from "@aura-stack/router"
-import { createAuthorizationURL } from "@/actions/signIn/authorization-url.ts"
-import { createOIDCAuthorizationURL } from "@/actions/oidc/authorization-url.ts"
-import { isOIDCProvider, resolveOpenIDProvider } from "@/actions/oidc/resolve-provider.ts"
-import { createRedirectTo, createRedirectURI, createSignInURL } from "@/actions/signIn/authorization.ts"
+import { cacheControl, secureApiHeaders } from "@/shared/headers.ts"
 import { createValidation, handleApiError } from "@/shared/utils/api.ts"
+import { createAuthorizationURL } from "@/actions/signIn/authorization-url.ts"
+import { createOIDCAuthorizationURL } from "@/shared/oidc/authorization-url.ts"
+import { isOIDCProvider, resolveOpenIDProvider } from "@/shared/oidc/resolve-provider.ts"
+import { createRedirectTo, createRedirectURI, createSignInURL } from "@/actions/signIn/authorization.ts"
 import type { BuiltInOAuthProvider, FunctionAPIContext, LiteralUnion, SignInAPIOptions, SignInAPIReturn } from "@/@types/index.ts"
 
 /**
