+++ packages/core/src/shared/oidc/discovery.ts
