+++ packages/core/src/shared/oidc/index.ts
@@ -1,0 +1,10 @@
+export { discoveryMetadata, DISCOVERY_SUFFIX, normalizeIssuer } from "@/shared/oidc/discovery.ts"
+export {
+    resolveOpenIDProvider,
+    isOIDCProvider,
+    createOpenIDPlaceholder,
+    clearResolvedProviderCache,
+} from "@/shared/oidc/resolve-provider.ts"
+export { fetchJWKS, getJWKSVerifier, ensureJWKSValidated, clearJWKSCache } from "@/shared/oidc/jwks.ts"
+export { validateIDToken, type ValidateIDTokenOptions } from "@/shared/oidc/id-token.ts"
+export { createOIDCAuthorizationURL } from "@/shared/oidc/authorization-url.ts"
