+++ packages/core/test/actions/callback/callback.test.ts
@@ -1,7 +1,7 @@
 import { describe, test, expect, vi, beforeEach, afterEach } from "vitest"
 import { GET, jose, oauthCustomService, openIDCustomProvider, openIDMetadata, RS256PEMFormat } from "@test/presets.ts"
-import { clearResolvedProviderCache } from "@/actions/oidc/resolve-provider.ts"
-import { clearJWKSCache } from "@/actions/oidc/jwks.ts"
+import { clearResolvedProviderCache } from "@/shared/oidc/resolve-provider.ts"
+import { clearJWKSCache } from "@/shared/oidc/jwks.ts"
 import { SignJWT, exportJWK, importPKCS8, importSPKI } from "@aura-stack/jose/jose"
 import { createPKCE } from "@/shared/crypto.ts"
 import { setCookie, getSetCookie } from "@/cookie.ts"
