+++ packages/core/test/actions/oidc/id-token.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect, vi, beforeEach, afterEach } from "vitest"
-import { validateIDToken } from "@/actions/oidc/id-token.ts"
-import { clearJWKSCache } from "@/actions/oidc/jwks.ts"
+import { validateIDToken } from "@/shared/oidc/id-token.ts"
+import { clearJWKSCache } from "@/shared/oidc/jwks.ts"
 import { RS256PEMFormat } from "@test/presets.ts"
 import { SignJWT, exportJWK, importPKCS8, importSPKI } from "@aura-stack/jose/jose"
